authorization

+++ routes/index.js
@@ -1,10 +1,16 @@
 var  express =require( 'express');
-var router = express.Router();
+var bcrypt = require('bcrypt');
+var ExpressJwt = require('express-jwt');
+var JWT = require('jsonwebtoken');
 var mongoose = require('mongoose');
 const Client = require('../db/model/Clients');
 var DBMethod = require('../db/db.metod');
 
+var router = express.Router();
 
+const saltRounds = 10;
+
+var Authorization = DBMethod.Authorization;
 var GetClients = DBMethod.GetClients;
 var GetClient = DBMethod.GetClient;
 var EditClient = DBMethod.EditClient;
@@ -12,6 +18,8 @@ var DeleteClient = DBMethod.DeletClient;
 var CreateClient = DBMethod.CreateClient;
 var FilterClient = DBMethod.FilterClient;
 var CreateUser = DBMethod.CreateUser;
+
+const publicKey = 'secret';
 
 router.use(express.json());
 
@@ -27,15 +35,27 @@ router.get('/users', (req, res) => {
         .then(respons => { res.json(respons) });    
 });
 
+router.use(ExpressJwt({ secret: publicKey}).unless({path: ['/users', /\/user\/.*/, /\/search\/.*/, '/signin', '/signup']}));
+
 router.post('/signin', (req, res) => {
-    console.log(req.body)
-    res.send();
+    Authorization(req.body.email)
+       .then(function(respons){
+           bcrypt.compare(req.body.password, respons.password).then(function(resp) {
+                resp ? res.status(200).json(respons): res.status(401).send();
+           });
+       });
 });
 
 router.post('/signup', (req, res) => {
-    console.log(req.body);
-    CreateUser(req.body)
-        .then(respons => { res.json(respons); });
+    bcrypt.hash(req.body.password, saltRounds).then(function(hash) {
+        req.body.password = hash;
+        req.body.role = 1;
+        JWT.sign({role:req.body.role, email: req.body.email}, publickKey, (err,token) => {
+            req.body.token = token;
+            CreateUser(req.body)
+                .then(respons => { res.json(respons); });
+        });
+    });
 });
 
 router.get('/user/:id', (req, res) => {
@@ -44,6 +64,7 @@ router.get('/user/:id', (req, res) => {
 });
 
 router.get('/search/:id', (req, res) => {
+    console.log(req.params.id);
     FilterClient(req.params.id)
         .then(respons => { res.json(respons) });
 });
@@ -60,8 +81,12 @@ router.post('/create', (req, res) => {
 
 
 router.get('/delete/:id', (req, res) => {
-    DeleteClient(req.params.id)
-        .then(respons => { res.send() });
+    if(req.user.role === 2){
+        DeleteClient(req.params.id)
+            .then(respons => { res.status(200).send() });
+    }else{
+        res.status(403).send();
+    }
 });
 
 module.exports = router;
